+++ src/graphql/routes/journal/schema.gql.ts
@@ -1,9 +1,13 @@
 import { SchemaFragment } from "../../types/schema.types";
 
 export default {
+    // 'id' is type BigInt in database
+    //    This returns with the postfix 'n', which GraphQL treats as a string
+    //    So convert BigInt to string in js code
+    //    and keep it as BigInt in Database
     Types: `
     type Journal {
-      id: Int!
+      id: BigInt!
       userId: String!
       name: String!
     }
